Makefile: allow building klib without building kernel
With the current code, trying to build the gpu_nvidia klib without
first building the kernel fails because the debug.h header file is
not generated automatically.
This issue is fixed by amending kernel_open/Makefile so that it
does not override $(GENHEADERS) (which would remove the debug.h
header file from this variable) and appends closure_templates.h to
this variable instead. In addition, in order to avoid creating
closure_templates.h outside $(OUTPUTDIR), $(OBJDIR) is overridden
before including rules.mk so that a
$(OUTPUTDIR)/closure_templates.h make rule is picked up (instead of
$(CURDIR)/closure_templates.h) from rules.mk.