uvm_nanos.h: move inclusion of net_system_structs.h
uvm_nanos.h includes the net_system_structs.h header file from the
kernel source because the UVM driver uses the standard Unix error
number definitions, some of which are in that header file. Since
the addition of the tcp_info struct definition in the header file,
building the UVM driver fails because of missing definitions of the
u8, u32 and u64 data types. Moving the inclusion of the header in
uvm_nanos.h to after the inclusion of unix_internal.h fixes the
build failure.